refactor: Player에서 rule에 대한 의존성 줄임
