+++ src/main/kotlin/omock/model/Player.kt
@@ -1,7 +1,11 @@
 package omock.model
 
+import omock.model.rule.OMockRule
+
 sealed class Player {
     abstract val stoneHistory: ArrayDeque<Stone>
+
+    abstract val oMockRule: OMockRule
 
     fun turn(action: () -> Pair<String, String>): Stone {
         val coordinate = action()
@@ -18,15 +22,4 @@ sealed class Player {
     }
 
     abstract fun judgementResult(visited: Map<Direction, Result>): Boolean
-
-    companion object {
-        const val INIT_COUNT = 0
-        const val MIN_FOUR_TO_FOUR_COUNT = 4
-        const val MIN_IS_CLEAR_FOUR_TO_FOUR_COUNT = 4
-        const val MIN_THREE_TO_THREE_COUNT = 4
-        const val MIN_REVERSE_COUNT = 0
-        const val MIN_O_MOCK_COUNT = 4
-        const val EDGE_THREE_TO_THREE_COUNT = 2
-        const val EDGE_FOUR_TO_FOUR_COUNT = 3
-    }
 }
