+++ src/main/kotlin/omock/model/WhitePlayer.kt
@@ -1,13 +1,13 @@
 package omock.model
 
-data class WhitePlayer(
+import omock.model.rule.OMockRule
+import omock.model.rule.WhiteRule
+
+class WhitePlayer(
     override val stoneHistory: ArrayDeque<Stone> = ArrayDeque(),
+    override val oMockRule: OMockRule = WhiteRule(),
 ) : Player() {
     override fun judgementResult(visited: Map<Direction, Result>): Boolean {
-        visited.entries.forEach { (key, result) ->
-            val reverseResultCount: Int = visited[Direction.reverse(key)]?.count ?: MIN_REVERSE_COUNT
-            return reverseResultCount + result.count >= MIN_O_MOCK_COUNT
-        }
-        return false
+        return oMockRule.judgementResult(visited)
     }
 }
