+++ src/main/kotlin/omock/model/rule/OMockRule.kt
@@ -1,0 +1,19 @@
+package omock.model.rule
+
+import omock.model.Direction
+import omock.model.Result
+
+interface OMockRule {
+    fun judgementResult(visited: Map<Direction, Result>): Boolean
+
+    companion object {
+        const val INIT_COUNT = 0
+        const val MIN_FOUR_TO_FOUR_COUNT = 4
+        const val MIN_IS_CLEAR_FOUR_TO_FOUR_COUNT = 4
+        const val MIN_THREE_TO_THREE_COUNT = 4
+        const val MIN_REVERSE_COUNT = 0
+        const val MIN_O_MOCK_COUNT = 4
+        const val EDGE_THREE_TO_THREE_COUNT = 2
+        const val EDGE_FOUR_TO_FOUR_COUNT = 3
+    }
+}
